chore: Switch integ tests to environment opt-out pattern
Run all integ tests in both node & browser, using file naming to designate tests that require a specific environment. This effectively switches us from environment-opt-in (need to duplicate tests between X.test and X.browser.test to have common coverage) to opt-out (need to explicitly move tests to x.test.node if it's only supposed to run in node).

Goal here is to reduce test duplication and improve compatibility between browser and node.
