+++ test/integ/bash.test.node.ts
