--- test/integ/browser/agent.browser.test.ts
@@ -1,119 +1,0 @@
-import { describe, it, expect } from 'vitest'
-import { Agent, DocumentBlock, ImageBlock, Message, TextBlock, tool } from '@strands-agents/sdk'
-import { notebook } from '@strands-agents/sdk/vended_tools/notebook'
-import { httpRequest } from '@strands-agents/sdk/vended_tools/http_request'
-import { z } from 'zod'
-
-import { collectGenerator } from '$/sdk/__fixtures__/model-test-helpers.js'
-
-// Import fixtures
-import yellowPngUrl from '../__resources__/yellow.png?url'
-import { loadFixture } from '../__fixtures__/test-helpers.js'
-import { allProviders } from '../__fixtures__/model-providers.js'
-
-// Calculator tool for testing
-const calculatorTool = tool({
-  name: 'calculator',
-  description: 'Performs basic arithmetic operations',
-  inputSchema: z.object({
-    operation: z.enum(['add', 'subtract', 'multiply', 'divide']),
-    a: z.number(),
-    b: z.number(),
-  }),
-  callback: async ({ operation, a, b }) => {
-    const ops = {
-      add: a + b,
-      subtract: a - b,
-      multiply: a * b,
-      divide: a / b,
-    }
-    return `Result: ${ops[operation]}`
-  },
-})
-
-describe.each(allProviders)('Agent Browser Tests with $name', async ({ name, skip, createModel }) => {
-  describe.skipIf(skip)(`${name} Browser Integration`, () => {
-    it('handles basic invocation', async () => {
-      const agent = new Agent({ model: createModel(), printer: false })
-      const result = await agent.invoke('Say hello in one word')
-
-      expect(result.stopReason).toBe('endTurn')
-      expect(result.lastMessage.role).toBe('assistant')
-      expect(result.lastMessage.content.length).toBeGreaterThan(0)
-    })
-
-    it('handles tool use', async () => {
-      const agent = new Agent({
-        model: createModel(),
-        printer: false,
-        systemPrompt: 'Use the calculator tool to solve math problems. Respond with only the numeric result.',
-        tools: [calculatorTool],
-      })
-
-      const { result } = await collectGenerator(agent.stream('What is 123 * 456?'))
-
-      // Verify tool was used
-      const toolUseMessage = agent.messages.find((msg) => msg.content.some((block) => block.type === 'toolUseBlock'))
-      expect(toolUseMessage).toBeDefined()
-
-      // Verify final response
-      expect(result.stopReason).toBe('endTurn')
-      expect(result.lastMessage.role).toBe('assistant')
-    })
-
-    it('handles media blocks', async () => {
-      const docBlock = new DocumentBlock({
-        name: 'test-document',
-        format: 'txt',
-        source: { text: 'The document contains the word ZEBRA.' },
-      })
-
-      const imageBytes = await loadFixture(yellowPngUrl)
-      const imageBlock = new ImageBlock({
-        format: 'png',
-        source: { bytes: imageBytes },
-      })
-
-      const agent = new Agent({
-        model: createModel(),
-        messages: [
-          new Message({
-            role: 'user',
-            content: [
-              docBlock,
-              imageBlock,
-              new TextBlock('What animal is in the document and what color is the image? Answer briefly.'),
-            ],
-          }),
-        ],
-        printer: false,
-      })
-
-      const result = await agent.invoke('Answer the question!')
-
-      expect(result.stopReason).toBe('endTurn')
-      expect(result.lastMessage.role).toBe('assistant')
-    })
-  })
-
-  it('handles tool invocation', async () => {
-    const agent = new Agent({
-      model: createModel(),
-      tools: [notebook, httpRequest],
-      printer: false,
-    })
-
-    await agent.invoke('Call Open-Meteo to get the weather in NYC, and take a note of it')
-
-    expect(
-      agent.messages.some((message) =>
-        message.content.some((block) => block.type == 'toolUseBlock' && block.name == 'notebook')
-      )
-    ).toBe(true)
-    expect(
-      agent.messages.some((message) =>
-        message.content.some((block) => block.type == 'toolUseBlock' && block.name == 'http_request')
-      )
-    ).toBe(true)
-  })
-})
