+++ test/integ/environment.test.ts
@@ -1,23 +1,6 @@
 import { describe, it, expect } from 'vitest'
 
-import { isNode } from '$/sdk/__fixtures__/environment.js'
-
 describe('environment', () => {
-  describe('Node.js compatibility', () => {
-    it('works in Node.js environment', () => {
-      // Test Node.js specific features are available
-      expect(typeof process).toBe('object')
-      expect(process.version).toBeDefined()
-    })
-  })
-
-  describe('environment detection', () => {
-    it('correctly identifies Node.js environment', () => {
-      expect(isNode).toBe(true)
-      expect(typeof process).toBe('object')
-    })
-  })
-
   describe('JavaScript features', () => {
     it('supports modern JavaScript features', () => {
       // Test ES2022 features work
