+++ test/integ/file-editor.test.node.ts
